Handle launching of boundscheck tests in a separate file.
Follows style of parallel tests.
